Added help page for mobile number setting
Added a help page for mobile number setting.
This is necessary for SMS channel. When the
number is not set instead of the normal SMS
icon there is a clickable SMS icon with a
warning sign. Clicking on the icon opens the
help page with instructions on how to set
the missing mobile phone number.

Added a tooltips on all icons.