Modified SMS class to be empty class
Modified SMS class to be an empty class
because the implementation of the SMS channel
depends on the provider settings. Provided an
SMS.php.sample class which can be used as sample.
Implemented a mechanism that disables SMS
functionality if the SMS class is not implemented.
Updated README.txt file with instruction on how to
proceede to enable SMS channel.