Removed SMS class and provided a SMS sample class
Replaced SMS class with a sample SMS class
because the implementation of the SMS channel
depends on the provider settings.
SMS.php.sample class which can be used as starting point.
Implemented a mechanism that disables SMS
functionality if the SMS class is not implemented.
Updated README.txt file with instruction on how to
proceede to enable SMS channel.