Conditional activities support fixed
The conditional activities were not managed correctly.
Now all the conditional activities which are set to be
invisible until the condition is matched are ignored.
Once the condition is matched the activity/resource is
treated as new and is notified through available channels.

BUGFIX: html tag span not closed correctly